+++ rocketgram/api/sticker_set.py
@@ -3,11 +3,13 @@
 # Rocketgram is released under the MIT License (see LICENSE).
 
 
+import warnings
 from dataclasses import dataclass
 from typing import List, Optional
 
 from .photo_size import PhotoSize
 from .sticker import Sticker
+from .sticker_type import StickerType
 
 
 @dataclass(frozen=True)
@@ -19,11 +21,16 @@ class StickerSet:
 
     name: str
     title: str
+    sticker_type: StickerType
     is_animated: Optional[bool]
     is_video: Optional[bool]
-    contains_masks: bool
     stickers: List[Sticker]
     thumb: Optional[PhotoSize]
+
+    @property
+    def contains_masks(self) -> bool:
+        warnings.warn("This field is deprecated.", DeprecationWarning)
+        return self.sticker_type is StickerType.mask
 
     @classmethod
     def parse(cls, data: dict) -> Optional['StickerSet']:
@@ -32,5 +39,10 @@ class StickerSet:
 
         stickers = [Sticker.parse(s) for s in data['stickers']]
 
-        return cls(data['name'], data['title'], data['is_animated'], data['is_video'], data['contains_masks'], stickers,
+        try:
+            sticker_type = StickerType(data['sticker_type'])
+        except ValueError:
+            sticker_type = StickerType.unknown
+
+        return cls(data['name'], data['title'], sticker_type, data['is_animated'], data['is_video'], stickers,
                    PhotoSize.parse(data['thumb']))